+++ paymentsheet/src/main/java/com/stripe/android/link/ui/paymentmenthod/NativeLinkFormHelperFactory.kt
@@ -2,6 +2,7 @@ package com.stripe.android.link.ui.paymentmenthod
 
 import androidx.lifecycle.viewModelScope
 import com.stripe.android.link.injection.NativeLinkComponent
+import com.stripe.android.paymentsheet.DefaultFormDefinitionFactory
 import com.stripe.android.paymentsheet.DefaultFormHelper
 import com.stripe.android.paymentsheet.FormHelper
 import com.stripe.android.paymentsheet.LinkInlineHandler
@@ -15,24 +16,31 @@ internal class NativeLinkFormHelperFactory(
     private val parentComponent: NativeLinkComponent,
 ) {
     fun create(): FormHelper {
+        val linkInlineHandler = LinkInlineHandler.create()
+        val paymentMethodMetadata = parentComponent.paymentMethodMetadata
+
         return DefaultFormHelper(
             coroutineScope = parentComponent.viewModel.viewModelScope,
-            linkInlineHandler = LinkInlineHandler.create(),
-            cardAccountRangeRepositoryFactory = parentComponent.cardAccountRangeRepositoryFactory,
-            paymentMethodMetadata = parentComponent.paymentMethodMetadata,
-            newPaymentSelectionProvider = { null },
-            linkConfigurationCoordinator = null,
+            linkInlineHandler = linkInlineHandler,
+            paymentMethodMetadata = paymentMethodMetadata,
             selectionUpdater = {},
-            setAsDefaultMatchesSaveForFutureUse =
-                FORM_ELEMENT_SET_DEFAULT_MATCHES_SAVE_FOR_FUTURE_DEFAULT_VALUE,
             eventReporter = parentComponent.eventReporter,
             savedStateHandle = parentComponent.viewModel.savedStateHandle,
-            autocompleteAddressInteractorFactory = createAutocompleteAddressInteractorFactory(),
-            isLinkUI = true,
-            automaticallyLaunchedCardScanFormDataHelper = null,
-            tapToAddHelper = null,
-            paymentMethodMessagePromotionsHelper = null,
-            isNfcScanningAvailable = null,
+            formDefinitionFactory = DefaultFormDefinitionFactory(
+                linkInlineHandler = linkInlineHandler,
+                cardAccountRangeRepositoryFactory = parentComponent.cardAccountRangeRepositoryFactory,
+                paymentMethodMetadata = paymentMethodMetadata,
+                newPaymentSelectionProvider = { null },
+                linkConfigurationCoordinator = null,
+                setAsDefaultMatchesSaveForFutureUse =
+                    FORM_ELEMENT_SET_DEFAULT_MATCHES_SAVE_FOR_FUTURE_DEFAULT_VALUE,
+                autocompleteAddressInteractorFactory = createAutocompleteAddressInteractorFactory(),
+                isLinkUI = true,
+                automaticallyLaunchedCardScanFormDataHelper = null,
+                tapToAddHelper = null,
+                paymentMethodMessagePromotionsHelper = null,
+                isNfcScanningAvailable = null,
+            ),
         )
     }
