+++ paymentsheet/src/main/java/com/stripe/android/paymentsheet/DefaultFormHelper.kt
@@ -1,57 +1,32 @@
 package com.stripe.android.paymentsheet
 
 import androidx.lifecycle.SavedStateHandle
-import com.stripe.android.cards.CardAccountRangeRepository
-import com.stripe.android.common.nfcscan.IsNfcScanningAvailable
-import com.stripe.android.common.taptoadd.TapToAddHelper
-import com.stripe.android.link.LinkConfigurationCoordinator
 import com.stripe.android.lpmfoundations.luxe.SupportedPaymentMethod
 import com.stripe.android.lpmfoundations.paymentmethod.PaymentMethodMetadata
-import com.stripe.android.lpmfoundations.paymentmethod.UiDefinitionFactory
-import com.stripe.android.model.PaymentMethod
 import com.stripe.android.model.PaymentMethodCode
-import com.stripe.android.model.PaymentMethodCreateParams
 import com.stripe.android.paymentsheet.FormHelper.FormType
 import com.stripe.android.paymentsheet.analytics.EventReporter
-import com.stripe.android.paymentsheet.forms.FormArgumentsFactory
 import com.stripe.android.paymentsheet.forms.FormFieldValues
 import com.stripe.android.paymentsheet.model.PaymentSelection
 import com.stripe.android.paymentsheet.model.paymentMethodType
-import com.stripe.android.paymentsheet.paymentdatacollection.FormArguments
-import com.stripe.android.paymentsheet.repositories.PaymentMethodMessagePromotionsHelper
-import com.stripe.android.paymentsheet.ui.transformToPaymentMethodCreateParams
 import com.stripe.android.paymentsheet.ui.transformToPaymentSelection
-import com.stripe.android.ui.core.elements.AutomaticallyLaunchedCardScanFormDataHelper
-import com.stripe.android.uicore.elements.AutocompleteAddressInteractor
-import com.stripe.android.uicore.elements.FormElement
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.combine
 import kotlinx.coroutines.launch
 
+private const val PREVIOUSLY_COMPLETED_PAYMENT_FORM = "previously_completed_payment_form"
+
 internal class DefaultFormHelper(
     private val coroutineScope: CoroutineScope,
     private val linkInlineHandler: LinkInlineHandler,
-    private val cardAccountRangeRepositoryFactory: CardAccountRangeRepository.Factory,
     private val paymentMethodMetadata: PaymentMethodMetadata,
-    private val newPaymentSelectionProvider: (PaymentMethodCode) -> NewPaymentOptionSelection?,
     private val selectionUpdater: (PaymentSelection?) -> Unit,
-    private val linkConfigurationCoordinator: LinkConfigurationCoordinator?,
-    private val setAsDefaultMatchesSaveForFutureUse: Boolean,
     private val eventReporter: EventReporter,
     private val savedStateHandle: SavedStateHandle,
-    private val autocompleteAddressInteractorFactory: AutocompleteAddressInteractor.Factory?,
-    private val isLinkUI: Boolean = false,
-    private val automaticallyLaunchedCardScanFormDataHelper: AutomaticallyLaunchedCardScanFormDataHelper?,
-    private val tapToAddHelper: TapToAddHelper?,
-    private val paymentMethodMessagePromotionsHelper: PaymentMethodMessagePromotionsHelper?,
-    private val isNfcScanningAvailable: IsNfcScanningAvailable?,
-) : FormHelper {
-    companion object {
-        internal const val PREVIOUSLY_COMPLETED_PAYMENT_FORM = "previously_completed_payment_form"
-    }
-
+    formDefinitionFactory: FormDefinitionFactory,
+) : FormHelper, FormDefinitionFactory by formDefinitionFactory {
     private val lastFormValues = MutableSharedFlow<Pair<FormFieldValues?, String>>(replay = 1)
 
     private val paymentSelection: Flow<PaymentSelection?> = combine(
@@ -80,56 +55,9 @@ internal class DefaultFormHelper(
         }
     }
 
-    override fun formElementsForCode(code: String): List<FormElement> {
-        return paymentMethodMetadata.formElementsForCode(
-            code = code,
-            uiDefinitionFactoryArgumentsFactory = createArgumentsFactory(code),
-        ) ?: emptyList()
-    }
-
-    override fun createFormArguments(
-        paymentMethodCode: PaymentMethodCode,
-    ): FormArguments {
-        return FormArgumentsFactory.create(
-            paymentMethodCode = paymentMethodCode,
-            metadata = paymentMethodMetadata,
-        )
-    }
-
     override fun onFormFieldValuesChanged(formValues: FormFieldValues?, selectedPaymentMethodCode: String) {
         coroutineScope.launch {
             lastFormValues.emit(formValues to selectedPaymentMethodCode)
-        }
-    }
-
-    override fun getPaymentMethodParams(
-        formValues: FormFieldValues?,
-        selectedPaymentMethodCode: String
-    ): PaymentMethodCreateParams? {
-        return formValues?.transformToPaymentMethodCreateParams(
-            paymentMethodCode = selectedPaymentMethodCode,
-            paymentMethodMetadata = paymentMethodMetadata
-        )
-    }
-
-    private fun requiresFormScreen(paymentMethodCode: String, formElements: List<FormElement>): Boolean {
-        val userInteractionAllowed = formElements.any { it.allowsUserInteraction }
-        return userInteractionAllowed ||
-            paymentMethodCode == PaymentMethod.Type.USBankAccount.code ||
-            paymentMethodCode == PaymentMethod.Type.Link.code
-    }
-
-    override fun formTypeForCode(paymentMethodCode: PaymentMethodCode): FormType {
-        val formElements = formElementsForCode(paymentMethodCode)
-        return if (requiresFormScreen(paymentMethodCode, formElements)) {
-            FormType.UserInteractionRequired
-        } else {
-            val mandate = formElements.firstNotNullOfOrNull { it.mandateText }
-            if (mandate == null) {
-                FormType.Empty
-            } else {
-                FormType.MandateOnly(mandate)
-            }
         }
     }
 
@@ -150,35 +78,5 @@ internal class DefaultFormHelper(
             eventReporter.onPaymentMethodFormCompleted(code)
             previouslyCompletedForm = code
         }
-    }
-
-    private fun createArgumentsFactory(code: String): UiDefinitionFactory.Arguments.Factory {
-        val currentSelection = newPaymentSelectionProvider(code)?.takeIf { it.getType() == code }
-
-        return UiDefinitionFactory.Arguments.Factory.Default(
-            cardAccountRangeRepositoryFactory = cardAccountRangeRepositoryFactory,
-            linkConfigurationCoordinator = linkConfigurationCoordinator,
-            linkInlineHandler = linkInlineHandler,
-            onLinkInlineSignupStateChanged = linkInlineHandler::onStateUpdated,
-            paymentMethodCreateParams = currentSelection?.getPaymentMethodCreateParams(),
-            paymentMethodOptionsParams = currentSelection?.getPaymentMethodOptionParams(),
-            paymentMethodExtraParams = currentSelection?.getPaymentMethodExtraParams(),
-            initialLinkUserInput = when (val selection = currentSelection?.paymentSelection) {
-                is PaymentSelection.New.Card -> selection.linkInput
-                else -> null
-            },
-            previousLinkSignupCheckboxSelection = when (val selection = currentSelection?.paymentSelection) {
-                // User entered a card and may have link input
-                is PaymentSelection.New.Card -> selection.linkInput != null
-                else -> null // Not a card, so no previous choice
-            },
-            setAsDefaultMatchesSaveForFutureUse = setAsDefaultMatchesSaveForFutureUse,
-            autocompleteAddressInteractorFactory = autocompleteAddressInteractorFactory,
-            isLinkUI = isLinkUI,
-            automaticallyLaunchedCardScanFormDataHelper = automaticallyLaunchedCardScanFormDataHelper,
-            tapToAddHelper = tapToAddHelper,
-            paymentMethodMessagingPromotionsHelper = paymentMethodMessagePromotionsHelper,
-            isNfcScanningAvailable = isNfcScanningAvailable,
-        )
     }
 }
